backend_doc/requirements.txt
aiohttp
aiosignal
alabaster
anyio
appdirs
argon2-cffi
argon2-cffi-bindings
arrow
asgiref
astroid
astropy
asttokens
async-timeout
atomicwrites
attrs
autobahn
Automat
autopep8
Babel
backcall
backports.functools-lru-cache
backports.tempfile
backports.weakref
backports.zoneinfo
bcrypt
beautifulsoup4
bidict
billiard
binaryornot
bitarray
bkcharts
black
bleach
bokeh
boto3
botocore
Bottleneck
brotlipy
CacheControl
cachetools
celery
certifi
cffi
chardet
charset-normalizer
click
click-didyoumean
click-plugins
click-repl
cloudpickle
colorama
colorcet
comtypes
constantly
cookiecutter
coreapi
coreschema
cryptography
cssselect
cycler
Cython
cytoolz
dask
debugpy
decorator
defusedxml
Deprecated
diff-match-patch
distlib
distributed
dj-database-url
Django
django-cors-headers
django-extensions
django-filter
django-multiselectfield
django-mysql
django-redis
django-session-timeout
django-storages
django-user-agents
djangorestframework
djangorestframework-csv
djangorestframework-simplejwt
dnspython
docopt
docutils
drf-extensions
drf-yasg
entrypoints
et-xmlfile
executing
fastjsonschema
filelock
firebase-admin
flake8
Flask
Flask-SocketIO
flower
fonttools
frozenlist
fsspec
future
fuzzywuzzy
gensim
glob2
google-api-core
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib
google-cloud-core
google-cloud-firestore
google-cloud-storage
google-crc32c
google-resumable-media
googleapis-common-protos
greenlet
grpcio
gspread
gunicorn
h5py
hashids
HeapDict
heroku
holoviews
httplib2
humanize
hvplot
hyperlink
idna
imagecodecs
imageio
imagesize
importlib-metadata
incremental
inflection
iniconfig
intake
intervaltree
ipykernel
ipython
ipython-genutils
ipywidgets
isort
itemadapter
itemloaders
itsdangerous
itypes
jdcal
jedi
Jinja2
jinja2-time
jmespath
joblib
json5
jsonpointer
jsonschema
jupyter
jupyter-client
jupyter-console
jupyter-core
jupyter-server
jupyter-server-proxy
jupyterlab
jupyterlab-pygments
jupyterlab-server
jupyterlab-widgets
keyring
kiwisolver
kombu
lazy-object-proxy
libarchive-c
llvmlite
locket
lxml
Markdown
MarkupSafe
matplotlib
matplotlib-inline
mistune
mkl-fft
mkl-random
mkl-service
mock
mpmath
msgpack
multidict
multipledispatch
munkres
mypy-extensions
mysql-connector-python
mysqlclient
nbclassic
nbclient
nbconvert
nbformat
nest-asyncio
networkx
nose
notebook
num2words
numba
numexpr
numpy
numpydoc
oauth2client
oauthlib
olefile
openpyxl
packaging
pandas
pandocfilters
panel
param
paramiko
parsel
parso
partd
pathspec
patsy
pdfkit
pep8
pexpect
pickleshare
Pillow
pip
pip-chill
pipreqs
pkginfo
platformdirs
plotly
pluggy
poyo
prometheus-client
prompt-toolkit
Protego
proto-plus
protobuf
psutil
psycopg2
psycopg2-binary
ptyprocess
pure-eval
py
pyasn1
pyasn1-modules
pycodestyle
pycosat
pycparser
pyct
pycurl
PyDispatcher
pydocstyle
pyerfa
pyflakes
Pygments
PyHamcrest
PyJWT
pylint
pylint-django
pylint-plugin-utils
pyls-spyder
PyMySQL
PyNaCl
pyodbc
pyOpenSSL
pyotp
pyparsing
PyPDF2
pyreadline
pyrsistent
PySocks
pytest
python-dateutil
python-engineio
python-Levenshtein
python-lsp-black
python-lsp-jsonrpc
python-lsp-server
python-magic
python-slugify
python-snappy
python-socketio
pytz
pyviz-comms
PyWavelets
PyYAML
pyzmq
QDarkStyle
qstylizer
QtAwesome
qtconsole
QtPy
queuelib
recordclass
redis
regex
reportlab
requests
requests-file
requests-oauthlib
requests-pkcs12
rope
rsa
Rtree
ruamel.yaml
ruamel.yaml.clib
s3transfer
scikit-image
scikit-learn
scipy
Scrapy
seaborn
Send2Trash
sentry-sdk
service-identity
setuptools
simpervisor
simplejson
six
smart-open
sniffio
snowballstemmer
sortedcollections
sortedcontainers
soupsieve
Sphinx
sphinxcontrib-applehelp
sphinxcontrib-devhelp
sphinxcontrib-htmlhelp
sphinxcontrib-jsmath
sphinxcontrib-qthelp
sphinxcontrib-serializinghtml
spyder
spyder-kernels
SQLAlchemy
sqlparse
stack-data
statsmodels
sympy
tables
tabulate
tblib
tenacity
terminado
testpath
text-unidecode
textdistance
threadpoolctl
three-merge
tifffile
tinycss
tldextract
toml
tomli
toolz
tornado
tqdm
traitlets
Twisted
txaio
typed-ast
typing_extensions
tzdata
ua-parser
ujson
unicodecsv
Unidecode
uritemplate
urllib3
user-agents
vine
virtualenv
vpython
w3lib
watchdog
wcwidth
webencodings
websocket-client
Werkzeug
wheel
whitenoise
widgetsnbextension
win-inet-pton
win-unicode-console
wincertstore
wrapt
xarray
xlrd
XlsxWriter
xmltodict
yapf
yarg
yarl
zict
